feat(storage): support any S3-compatible provider for session replays
Session replay batches were stored only in Cloudflare R2, with the
endpoint, region and credentials hardcoded to R2. The service already
used the AWS S3 SDK, so generalize it to any S3-compatible provider
(AWS S3, R2, MinIO, ...):

- Rename r2StorageService -> s3StorageService (export objectStorage) and
  update the session-replay ingest/query consumers.
- Configure via generic S3_* env vars (S3_ENDPOINT, S3_REGION,
  S3_ACCESS_KEY_ID, S3_SECRET_ACCESS_KEY, S3_BUCKET_NAME,
  S3_FORCE_PATH_STYLE). Enabled whenever credentials are present, so
  self-hosted instances can offload replays without IS_CLOUD.
- Keep legacy R2_* vars as a fallback and derive the R2 endpoint from
  R2_ACCOUNT_ID, so existing cloud/R2 deployments work unchanged.
- Document the new variables and wire them into docker-compose.yml.
